+++ src/main/java/com/ifgoiano/biblioteca/service/UsuarioService.java
@@ -5,8 +5,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.ifgoiano.biblioteca.model.Emprestimo;
 import com.ifgoiano.biblioteca.model.ResourceNotFoundException;
 import com.ifgoiano.biblioteca.model.Usuario;
+import com.ifgoiano.biblioteca.model.UsuarioComEmprestimosException;
+import com.ifgoiano.biblioteca.repository.EmprestimoRepository;
 import com.ifgoiano.biblioteca.repository.UsuarioRepository;
 
 // Criação do serviço de Usuário que encapsula a lógica de negócios
@@ -15,6 +18,9 @@ public class UsuarioService implements IUsuarioService {
 
     @Autowired
     private UsuarioRepository usuarioRepository;
+
+    @Autowired
+    private EmprestimoRepository emprestimoRepository;
 
     @Override
     public Usuario save(Usuario usuario) {
@@ -34,8 +40,12 @@ public class UsuarioService implements IUsuarioService {
 
     @Override
     public void deleteById(Long id) {
-        usuarioRepository.deleteById(id);
+    List<Emprestimo> emprestimos = emprestimoRepository.findByUsuarioId(id);
+    if (!emprestimos.isEmpty()) {
+        throw new UsuarioComEmprestimosException("Não é possível excluir o usuário, pois já houveram registros de empréstimo na Biblioteca.");
     }
+    usuarioRepository.deleteById(id);
+}
 
     @Override
     public Usuario authenticate(String login, String senha) {
